refactor: split AssetClassification.cs/MarketData.cs, move GlobalAssetClassMapping to Rules
Neither AssetClassification nor MarketData exist as types - both were
files named after nothing they actually defined, each bundling several
unrelated top-level types together (CountryCode/GlobalAssetClass/
GlobalAssetClassMapping; DividendType/AssetValueSnapshot/DividendValue).

Split each along type boundaries, matching the one-top-level-type-per-file
convention already used elsewhere (PositionType.cs). GlobalAssetClassMapping
is a pure static lookup with no state/IO, structurally identical to what
already lives in Financial.Domain/Rules (DividendValuationRules,
ProfitCalculator, XirrCalculator), so it joins that folder instead of
staying in Entities.
