+++ Financial.Domain/Entities/Asset.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Financial.Domain.Rules;
 
 namespace Financial.Domain.Entities;
