+++ Financial.Domain/Entities/CountryCode.cs
@@ -1,0 +1,9 @@
+namespace Financial.Domain.Entities;
+
+public enum CountryCode
+{
+    Unknown = 0,
+    BR,
+    US,
+    UK
+}
