+++ Financial.Domain/Entities/GlobalAssetClass.cs
@@ -1,0 +1,16 @@
+namespace Financial.Domain.Entities;
+
+public enum GlobalAssetClass
+{
+    Unknown = 0,
+    Equity,
+    RealEstate,
+    Bond,
+    Fund,
+    ETF,
+    Cash,
+    Pension,
+    Other,
+    Cryptocurrency,
+    PrivateCredit
+}
