+++ Financial.Domain/ValueObjects/DividendType.cs
@@ -1,0 +1,3 @@
+namespace Financial.Domain.ValueObjects;
+
+public enum DividendType { Dividend, JCP }
