+++ Financial.Domain/ValueObjects/DividendValue.cs
@@ -1,0 +1,5 @@
+using System;
+
+namespace Financial.Domain.ValueObjects;
+
+public record DividendValue(DividendType Type, DateTime Date, decimal Value);
